+++ unittesting/utils/__init__.py
@@ -2,7 +2,7 @@ from .json_file import JsonFile
 from .output_panel import OutputPanel
 from . import settings as UTSetting
 from .progress_bar import ProgressBar
-from .reload_package import reload_package
+from .reloader import reload_package
 
 __all__ = [
     "JsonFile",
